fix: improve script handling
